Cleaning code

+++ src/com/garrapeta/jumplings/game/JumplingsWorld.java
@@ -77,7 +77,7 @@ public class JumplingsWorld extends Box2DWorld {
     // ----------------------------------------------------------- Constructor
 
     public JumplingsWorld(FragmentActivity activity, GameView gameView, Context context) {
-        super(gameView, context, (short) (PermData.isSoundEnabled(context) ? PermData.CFG_LEVEL_ALL : PermData.CFG_LEVEL_NONE),
+        super(gameView, context, PermData.isSoundEnabled(context) ? PermData.CFG_LEVEL_ALL : PermData.CFG_LEVEL_NONE,
                 PermData.getVibratorLevel(context));
         mActivity = activity;
         mFactory = new JumplingsFactory(this);
